Clean up Storybook configuration
