+++ .storybook/theme.js
@@ -25,7 +25,7 @@ export default create({
   textInverseColor: 'rgb(255, 255, 255)',
 
   // Branding
-  brandTitle: 'Vocabulary',
+  brandTitle: 'Fonts',
   brandUrl: 'https://opensource.creativecommons.org/cc-fonts',
   brandImage: 'https://raw.githubusercontent.com/creativecommons/fonts/master/readme_assets/fonts_logo.svg?sanitize=true'
 })
