+++ package.json
@@ -16,7 +16,6 @@
     "@babel/core": "^7.8.3",
     "@storybook/addon-backgrounds": "^5.3.13",
     "@storybook/addon-docs": "^5.3.17",
-    "@storybook/addon-knobs": "^5.3.9",
     "@storybook/html": "^5.3.9",
     "babel-jest": "^25.2.3",
     "babel-loader": "^8.0.6",
